more repository code cleanup
