+++ bentoml/yatai/repository/s3_repository.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import logging
 from urllib.parse import urlparse
 
 import boto3
@@ -20,7 +21,9 @@ from botocore.exceptions import ClientError
 from bentoml import config
 from bentoml.exceptions import YataiRepositoryException
 from bentoml.yatai.proto.repository_pb2 import BentoUri
-from bentoml.yatai.repository.repository import BentoRepositoryBase, logger
+from bentoml.yatai.repository.base_repository import BentoRepositoryBase
+
+logger = logging.getLogger(__name__)
 
 
 class S3BentoRepository(BentoRepositoryBase):
